+++ packages/core/package.json
@@ -47,12 +47,12 @@
     "zod": "^3.11.6",
     "pino": "^7.8.0",
     "pino-pretty": "^7.5.1",
+    "vite": "2.9.1",
     "pino-abstract-transport": "^0.5.0",
     "@gestaltjs/vue-gestalt-renderer": "0.7.3",
     "@gestaltjs/renderer": "0.7.3"
   },
   "devDependencies": {
-    "@iarna/toml": "2.2.5",
     "@types/pino": "7.0.5",
     "@types/shelljs": "0.8.11",
     "case-anything": "2.1.10",
@@ -65,7 +65,6 @@
     "shelljs": "0.8.5",
     "tempy": "2.0.0",
     "terminal-link": "3.0.0",
-    "vite": "2.9.1",
     "zod": "3.14.3",
     "handlebars": "4.7.7"
   },
