--- packages/core/src/cli/app.ts
@@ -1,3 +1,0 @@
-export { default as load } from './app/load/app'
-export type { default as App } from './app/models/app'
-export type { default as Configuration } from '../shared/configuration'
