+++ packages/core/src/cli/constants.ts
@@ -1,10 +1,5 @@
-const constants = {
-  environmentVariables: {
-    /** Specify the Node environment the app is running in (production/development) */
-    environment: 'GESTALT_ENV',
-  },
+export const environmentVariables = {
+  environment: 'GESTALT_ENV',
 }
 
 export const configurationFileName = 'gestalt.config'
-
-export default constants
