+++ packages/core/src/cli/index.ts
@@ -7,7 +7,7 @@ export * as cli from './cli'
 export * as environment from './environment'
 export * as stream from './stream'
 export * as terminal from './terminal'
-export { default as constants } from './constants'
-export * as app from './app'
+export * as constants from './constants'
+export * as project from './project'
 export * as fs from './fs'
 export * as template from './template'
