+++ packages/core/src/cli/project/load/config.ts
