+++ packages/testing/src/workspace.ts
@@ -5,16 +5,31 @@ export function packagesDirectory(): string {
   return path.resolve(path.dirname(fileURLToPath(import.meta.url)), '../..')
 }
 
-type GestaltJSPackageModule = {
-  name: string
+/** It represents a ES module */
+type ESModule = {
+  /** Module identifier  */
+  identifier: string
+  /** Module path */
   path: string
 }
 
+/**
+ * An interface to represent the modules exported by the "gestaltjs" package.
+ */
 type GestaltJSPackageModules = {
-  support: GestaltJSPackageModule
-  configuration: GestaltJSPackageModule
+  /** Represents the support module which provides a set of convenient tools for building Gestalt projects. */
+  support: ESModule
+
+  /** Represents the configuration module which exports the defineConfiguration module for defining a configuration. */
+  configuration: ESModule
 }
 
+/**
+ * This function returns the modules exported by the "gestaltjs" NPM package.
+ * This is useful for tests that invoke a build system like Vite and need to instruct it on
+ * how to resolve those modules.
+ * @returns {GestaltJSPackageModules} An object representing the Node modules that are exported by the "gestaltjs" NPM package.
+ */
 export function gestaltjsPackageModules(): GestaltJSPackageModules {
   const frameworkDirectory = path.resolve(
     path.dirname(fileURLToPath(import.meta.url)),
@@ -22,11 +37,11 @@ export function gestaltjsPackageModules(): GestaltJSPackageModules {
   )
   return {
     support: {
-      name: 'gestaltjs/support',
+      identifier: 'gestaltjs/support',
       path: path.join(frameworkDirectory, 'support.ts'),
     },
     configuration: {
-      name: 'gestaltjs/configuration',
+      identifier: 'gestaltjs/configuration',
       path: path.join(frameworkDirectory, 'configuration.ts'),
     },
   }
